Fix default memory_limiter configuration for helm version 3.5.x
Starting from helm 3.5.0, merging rules were changed: now if null value merged into a dict, null value is preserved.

In order to make it cleaner and working on all the versions, this commit replaces the merge operation with a conditional set.

If user wants to override the default memory_limiter configuration, they should override all the fields. That behavior seems more practical than partial override for the this processor.