[opentelemetry-collector] Add extra ingress configurations (#311)
This commit adds a new values parameter `additionalIngresses` to the OpenTelemetry Collector Helm chart. Use case for this is where multiple receivers are deployed in a single OpenTelemetry Collector configuration that need different ingress annotations/configuration. For example:

- Receiver A (OTLP receiver etc.) uses NGINX ingress controller but needs to whitelist IPs to client CIDR ranges (with annotations) to only allow metrics sent from external client environments

- Receiver B (AWS CloudWatch receiver etc.) uses NGINX ingress controller but needs to whitelist AWS CIDR ranges (with annotations) to only allow metrics sent from AWS platform through CloudWatch Metric Streams
This can currently be acomplished by manually deploying another ingress resource or by deploying another Helm installation, but this change allows it to be done in a single installation instead.

Signed-off-by: John Clark <clark.john.d@outlook.com>