fix: Use port number instead of name in targetPort
When creating ports in a Service for a Deployment, the chart currently uses the 
port name (instead of port number) for the targetPort property value. This works 
in Kubernetes with Services and Ingresses, because it knows how to map port 
names to numbers.

However, third-party controllers, including in my case, the AWS Load Balancer 
Controller: https://kubernetes-sigs.github.io/aws-load-balancer-controller/v2.4/,
may not properly map port names to numbers when translating service ports into 
load balancer listener configurations. This means that this chart cannot create 
a functional Ingress using an AWS Application Load Balancer.

This fix is to simply use the port number for the service port. This shouldn't 
affect any existing Services created by the chart, because the port number 
works in Kubernetes equally as well as the name. It also shouldn't affect any 
externally created Ingresses that refer to the port, since the port name 
remains available and unchanged.