fix: add opentelemetry-operator.labels template to deployment
This commit moves the labels that were in the main template for the operator deployment manifest file into the existing named template in the _helpers.tpl file. It then uses that named template in the original deployment file. This fixes an issue where the operator chart would create a deployment object that didn't follow helm best practices.