[demo] Support overriding of env variables (#484)
Enables #438 by adding new properties to override environment variables. It supports to overwrite and/or add variables to the default section or to every individual component without the need of repeating the whole env lists.

It indirectly also adds support for templated default envs, so using templates in the env variable of the default section.

It adds only new properties and with that the approach is fully compatible.

The merge logic for the environment variables is implemented in a dedicated helper template which gets applied twice; for generating the default envs and the component specific ones.

The PR does not support to override in the component an environment variable defined in the default section. That will require to print the env variables all together (default and component specific) and would not allow to have a re-usable template for the merge logic.
You would need a way to first merge the default envs, then append the result of that to the component env list, then merge that with the component overrides and then print it all together. That is possible and worked out but had repeating logic and the overall template looked very complex. So I decided for now to take a more clean approach but not having the capability supported.