[demo] use Jaeger sub-chart (#523)
Moves to using a Jaeger sub-chart instead of templated within this chart itself.

This does introduce a potential breaking change, where the Jaeger parameters are moving to a top-level jaeger. section instead. Even though this is a breaking change, the chart is early enough, and the current section isn't likely to of been modified by existing deployments.

A readme is also added to the chart to document the upgrade process and all parameters in the chart itself.