[opentelemetry-demo] Adds liveness probe option to demo components (#616)
This allows components to be configured so that they can be
recreated if an endpoint fails to return a success code.

Specifically, this is configured for the featureflagService in
order to resolve:

https://github.com/open-telemetry/opentelemetry-demo/issues/695

When the ffspostgres pod gets recreated (for any reason) the database
is deleted and migrations (tables) are lost. This causes the
featureflag index endpoint to error. The liveness probe recreates the
featureflag pod, which runs its migrations, restoring a working state.