Add "app" labels for when integrated with Istio
I added explicit "app" labels to the Kubernetes Helm Deployment , as recommended by Istio. These labels add contextual information to the metrics and telemetry that Istio collects, which will help with visualization and analysis of the data.  Using these labels in conjunction with tools like Kiali and calisti can create nicer visualizations of the system topology, making it easier to understand and analyze the relationships between different components of our system.

ref:
https://istio.io/latest/docs/ops/deployment/requirements/