Bring back `configMap.existingPath` option
There doesn't seem to be a way to get at rendered template contents in Helm. This key allows a user to specify the path to template file that will be used to create `configMap.existingName`, in cases like mine where we want to wrap this chart in a chart that feeds in custom configmap contents.