[demo] Remove duplicate envvar on frontendproxy
OTEL_SERVICE_NAME envvar is set on all components as one of envvars listed in
default.env, to the value of app.kubernetes.io/component label. Recently
51993de76504 ("Bump dependencies and fix grafana dashboards") duplicated
OTEL_SERVICE_NAME on the frontendproxy deployment by setting it explicitly to
frontend-proxy. A duplicated envvar causes an error when upgrading the
opentelemetry-demo chart:

     Deployment.apps "opentelemetry-demo-frontendproxy" is invalid:
     spec.template.spec.containers[0].env[0].valueFrom: Invalid value: "": may
     not be specified when `value` is not empty

This commit removes the duplicated envvar.

Signed-off-by: Anna Kapuscinska <anna@isovalent.com>