[opentelemetry-integration] Bump collector version to v0.116.0 (#132)
Fixes ES-408.

To avoid confusion: the Collector version itself is 0.116.0, but apparently there was an issue with the Docker image for that version.

To solve the issue the OTEL team had to rebuild the image, which resulted in the image tag version diverging from the Collector binary version:

```
$ docker run --rm -it otel/opentelemetry-collector-contrib:0.116.1 -v
> otelcol-contrib version 0.116.0
```

Tested on an EKS cluster with both Windows and Linux nodes.