update schema to allow prestop sleep action in lifecycle hook
https://github.com/kubernetes/enhancements/blob/master/keps/sig-node/3960-pod-lifecycle-sleep-action/README.md

In order to avoid network errors during deployments or pod
interruptions, a pre stop sleep is often used to allow time for the
async endpoint removal process to complete. Kubernetes supports doing
the sleep natively now, removing the need for an exec based sleep
command. The values schema for the chart was blocking the
use of this feature.