[collector][supervisor] Fix custom image helper (#218)
Fixes ES-629.

This PR fixes an issue in the image helper. Before we had the default Collector image hardcoded in values so it would always be picked even when the Supervisor preset was enabled. Now by default there is no image hardcoded in the values. This allows the helper logic to work correctly:

If an image is given via values, no matter the circusmtances, always use it.
If no image is given via values and Supervisor preset is disabled: use the hardcoded default one.
If no image is given via values and Supervisor preset is enabled: use the custom supervised collector image.

Unfortunately this bug happened because I was using a custom image repository during my tests without realizing.