feat(collector): add `presets.coralogixExporter.pipelines` to allow enabling exporter on many pipelines at once (#234)
Fixes CDS-2254

This change will add new variable `presets.coralogixExporter.pipelines` as an `array` of `string` to allow enabling exporter on 2 pipelines at once.

So that in cases where we want to disable, let's say, coralogix exporter only in one pipeline, we could do this:

```
opentelemetry-agent:
  presets:
    coralogixExporter:
      enabled: true
      pipelines: ["metrics", "traces"]
```

instead of this:

```
opentelemetry-agent:
  presets:
    coralogixExporter:
      enabled: true
      pipeline: "none"
  config:
    service:
      pipelines:
        metrics:
          exporters:
            - coralogix
        traces:
          exporters:
            - coralogix
```

We've added a new variable instead of changing the existing one (`presets.coralogixExporter.pipeline`) for backward compatibility. The existing one `pipeline` will be deprecated in the future. For now, `pipelineS` will take precedence, while `pipeline` will be converted into array when set.