[feat]  Reduced attributes preset with support for logs and traces (#246)
Fixes ES-562.

This PR enhances the `reduceAttributes` preset with support for reducing the attributes in logs and traces besides metrics. A single processor is used for this, with statements for all signals. This processor is added to each one of the logs/metrics/traces pipelines that are present.

I added a new example to focus on this feature. It's available at `charts/opentelemetry-collector/examples/daemonset-reduced-attributes/`

I also added a few more attributes to the list of attributes we are dropping.