fix(collector): merge `resource` field into resource attributes (#275)
Fixes ES-716.

This should merge the `body.resource` field of the logs of our Collectors (all of them). into the resource attributes of the log entry. 

Currently it processes the logs of all the containers of our integration pods. This is the case because this configuration is in the `agent`, but it also collects logs from the `cluster-collector` and other agent types, and the container name is dynamic. 

A few examples to check: `daemonset-presets` and `daemonset-reduced-attributes`, among others.

Has been running safely in our online boutique demo cluster since a while and behaves as expected.