Add optional clusterRole.existingRoleName
Currently, it's either don't create a clusterRole, or create a new one.
There might be the case where one would like to just reuse an existing
cluster role. kube-static-metrics for instance already allows this.

With this change, the `opentelemetry-collector.clusterRoleName` is set
to the specified `clusterRole.existingName`; if not, just the same logic
as before.

The possible ambiguity of specifying both `clusterRole.useExistingRole`
and `create: true` by failing loud if both are set, instead of trying to
implicitly solving in some non obvious way.

Left it commented out since it's the less intrusive way of adding it,
which also means that the examples are not generated for this field.